+++ src/main/java/forif/univ_hanyang/user/repository/ForifTeamRepository.java
@@ -6,8 +6,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface ForifTeamRepository extends JpaRepository<ForifTeam, Integer> {
+public interface ForifTeamRepository extends JpaRepository<ForifTeam, Long> {
     Optional<List<ForifTeam>> findAllById_ActYearAndId_ActSemester(Integer actYear, Integer actSemester);
-    Optional<ForifTeam> findById_UserId(Integer id);
-    void deleteAllById_UserId(Integer userId);
+    Optional<ForifTeam> findById_UserId(Long id);
+    void deleteAllById_UserId(Long userId);
 }
